requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.24.0
numpy==1.26.0
sentence-transformers==2.2.2
python-dotenv==1.0.1
google-generativeai==0.14.0

